Add mirror to README (#667)
* Update README.md to add mirror

I added `mirror.mwt.me` as a possible repo option. I tried to keep this stylistically consistent with the rest of the document.

I also offloaded the GPG key download to keys.openpgp.org. This allows this step to work even if PackageCloud is out of bandwidth. It also has the advantage of putting the key id in the README and does not require users to trust me.

* undo use of openpgp.org

* Mention AppImage and Flatpak